build-tools/stx/patch/requirements.txt
click
lxml
pycryptodomex